+++ react/CheckoutB2B.tsx
@@ -14,6 +14,8 @@ import {
   withToast,
 } from 'vtex.styleguide'
 
+import 'vtex.country-codes/locales'
+
 import { CheckoutB2BProvider } from './CheckoutB2BContext'
 import { ContactInfos } from './components/ContactInfos'
 import {
@@ -44,9 +46,6 @@ function CheckoutB2B({ showToast }: WithToast) {
   const { items } = orderForm
   const loading = orderFormLoading || organizationLoading
   const filteredItems = toolbar?.filteredItems ?? items
-
-  // eslint-disable-next-line no-console
-  console.log('ORDER FORM:', orderForm)
 
   return (
     <div className={handles.container}>