+++ react/components/ContactInfos.tsx
@@ -5,7 +5,8 @@ import { Totalizer } from 'vtex.styleguide'
 import { useOrderFormCustom } from '../hooks'
 import type { CustomOrganization } from '../typings'
 import { messages } from '../utils'
-import { Address } from './Address'
+import { BillingAddress } from './BillingAddress'
+import { ShippingAddress } from './ShippingAddress'
 import { ShippingOption } from './ShippingOption'
 import { TruncatedText } from './TruncatedText'
 
@@ -83,7 +84,12 @@ export function ContactInfos({ organization }: Props) {
 
   contactFields.push({
     label: formatMessage(messages.shippingAddress),
-    value: <TruncatedText text={<Address />} />,
+    value: <TruncatedText text={<ShippingAddress />} />,
+  })
+
+  contactFields.push({
+    label: formatMessage(messages.billingAddress),
+    value: <BillingAddress />,
   })
 
   if (items.length) {