+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- New field for billing address.
 - Pending state for order placed button.
 
 ## [0.0.8] - 2024-11-12